🔧 Move Crew Process Management and Monitoring modules out of main loop file.
- Added `crew_process_starters.py` module to encapsulate functions for starting CrewManager and CrewRunner processes, addressing circular import issues.
- Implemented `CrewMonitor` class in `crew_monitor.py` for dynamic monitoring and management of Crew bots, including improved process handling and status checks.
- Enhanced logging capabilities for better traceability during bot operations.
- Updated `bots_main.py` to initialize and run the CrewMonitor, ensuring robust error handling and graceful shutdown procedures.
- Introduced periodic checks for bot status and management, improving overall system reliability.
